+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/truffle/test/BytecodeDSLCompilationTest.java
@@ -31,6 +31,7 @@ import static org.junit.Assert.assertNull;
 import static org.junit.Assert.fail;
 import static org.junit.Assume.assumeTrue;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.graalvm.polyglot.Context;
@@ -51,8 +52,10 @@ import com.oracle.truffle.api.bytecode.BytecodeRootNodes;
 import com.oracle.truffle.api.bytecode.ContinuationResult;
 import com.oracle.truffle.api.bytecode.test.BytecodeDSLTestLanguage;
 import com.oracle.truffle.api.bytecode.test.basic_interpreter.AbstractBasicInterpreterTest;
+import com.oracle.truffle.api.bytecode.test.basic_interpreter.AbstractBasicInterpreterTest.TestRun;
 import com.oracle.truffle.api.bytecode.test.basic_interpreter.BasicInterpreter;
 import com.oracle.truffle.api.bytecode.test.basic_interpreter.BasicInterpreterBuilder;
+import com.oracle.truffle.api.bytecode.test.basic_interpreter.BasicInterpreterBuilder.BytecodeVariant;
 import com.oracle.truffle.api.frame.FrameSlotKind;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.instrumentation.ExecutionEventNode;
@@ -69,14 +72,18 @@ import com.oracle.truffle.runtime.OptimizedCallTarget;
 public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
 
     @Parameters(name = "{0}")
-    public static List<Class<? extends BasicInterpreter>> getInterpreterClasses() {
-        return AbstractBasicInterpreterTest.allInterpreters();
+    public static List<TestRun> getParameters() {
+        List<TestRun> result = new ArrayList<>();
+        for (BytecodeVariant bc : AbstractBasicInterpreterTest.allVariants()) {
+            result.add(new TestRun(bc, false, false));
+        }
+        return result;
     }
 
-    @Parameter(0) public Class<? extends BasicInterpreter> interpreterClass;
+    @Parameter(0) public TestRun run;
 
     private boolean hasBoxingElimination() {
-        return new AbstractBasicInterpreterTest.TestRun(interpreterClass, false).hasBoxingElimination();
+        return run.hasBoxingElimination();
     }
 
     Context context;
@@ -123,7 +130,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testOSR1() {
-        BasicInterpreter root = parseNode(interpreterClass, BytecodeDSLTestLanguage.REF.get(null), false, "osrRoot", b -> {
+        BasicInterpreter root = parseNode(run, BytecodeDSLTestLanguage.REF.get(null), "osrRoot", b -> {
             b.beginRoot();
 
             BytecodeLocal iLoc = b.createLocal();
@@ -223,7 +230,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testOSR2() {
-        BasicInterpreter root = parseNode(interpreterClass, BytecodeDSLTestLanguage.REF.get(null), false, "osrRoot", b -> {
+        BasicInterpreter root = parseNode(run, BytecodeDSLTestLanguage.REF.get(null), "osrRoot", b -> {
             b.beginRoot();
 
             BytecodeLocal iLoc = b.createLocal();
@@ -341,7 +348,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
 
     @Test
     public void testCompiles() {
-        BasicInterpreter root = parseNodeForCompilation(interpreterClass, "addTwoConstants", b -> {
+        BasicInterpreter root = parseNodeForCompilation(run, "addTwoConstants", b -> {
             b.beginRoot();
 
             b.beginReturn();
@@ -365,7 +372,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
     @Test
     public void testMultipleReturns() {
         // return 30 + (arg0 ? 12 : (return 123; 0))
-        BasicInterpreter root = parseNodeForCompilation(interpreterClass, "multipleReturns", b -> {
+        BasicInterpreter root = parseNodeForCompilation(run, "multipleReturns", b -> {
             b.beginRoot();
 
             b.beginReturn();
@@ -403,7 +410,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
     @Test
     public void testStoreInvalidatesCode() {
         assumeTrue(hasBoxingElimination());
-        BytecodeRootNodes<BasicInterpreter> rootNodes = createNodes(interpreterClass, BytecodeDSLTestLanguage.REF.get(null), false, BytecodeConfig.DEFAULT, b -> {
+        BytecodeRootNodes<BasicInterpreter> rootNodes = createNodes(run, BytecodeDSLTestLanguage.REF.get(null), BytecodeConfig.DEFAULT, b -> {
             b.beginRoot();
             BytecodeLocal x = b.createLocal("x", null);
             b.beginStoreLocal(x);
@@ -463,7 +470,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
     @Test
     public void testBytecodeNodeStoreInvalidatesCode() {
         assumeTrue(hasBoxingElimination());
-        BytecodeRootNodes<BasicInterpreter> rootNodes = createNodes(interpreterClass, BytecodeDSLTestLanguage.REF.get(null), false, BytecodeConfig.DEFAULT, b -> {
+        BytecodeRootNodes<BasicInterpreter> rootNodes = createNodes(run, BytecodeDSLTestLanguage.REF.get(null), BytecodeConfig.DEFAULT, b -> {
             b.beginRoot();
             BytecodeLocal x = b.createLocal("x", null);
             b.beginStoreLocal(x);
@@ -537,7 +544,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
     @Test
     public void testMaterializedStoreInvalidatesCode() {
         assumeTrue(hasBoxingElimination());
-        BytecodeRootNodes<BasicInterpreter> rootNodes = createNodes(interpreterClass, BytecodeDSLTestLanguage.REF.get(null), false, BytecodeConfig.DEFAULT, b -> {
+        BytecodeRootNodes<BasicInterpreter> rootNodes = createNodes(run, BytecodeDSLTestLanguage.REF.get(null), BytecodeConfig.DEFAULT, b -> {
             b.beginRoot();
             BytecodeLocal x = b.createLocal("x", null);
             b.beginStoreLocal(x);
@@ -615,7 +622,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
     @Test
     public void testMaterializedAccessorStoreInvalidatesCode() {
         assumeTrue(hasBoxingElimination());
-        BytecodeRootNodes<BasicInterpreter> rootNodes = createNodes(interpreterClass, BytecodeDSLTestLanguage.REF.get(null), false, BytecodeConfig.DEFAULT, b -> {
+        BytecodeRootNodes<BasicInterpreter> rootNodes = createNodes(run, BytecodeDSLTestLanguage.REF.get(null), BytecodeConfig.DEFAULT, b -> {
             b.beginRoot();
             BytecodeLocal x = b.createLocal("x", null);
             b.beginStoreLocal(x);
@@ -688,7 +695,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
 
     @Test
     public void testInstrumentation() {
-        BasicInterpreter root = parseNodeForCompilation(interpreterClass, "addTwoConstantsInstrumented", b -> {
+        BasicInterpreter root = parseNodeForCompilation(run, "addTwoConstantsInstrumented", b -> {
             b.beginRoot();
 
             b.beginReturn();
@@ -710,7 +717,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
 
         // Instrumentation should invalidate the compiled code.
         root.getRootNodes().update(
-                        BasicInterpreterBuilder.invokeNewConfigBuilder(interpreterClass).addInstrumentation(BasicInterpreter.IncrementValue.class).build());
+                        run.bytecode().newConfigBuilder().addInstrumentation(BasicInterpreter.IncrementValue.class).build());
         assertNotCompiled(target);
 
         // The instrumented interpreter should be recompiled.
@@ -723,7 +730,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
 
     @Test
     public void testYield() {
-        BasicInterpreter root = parseNodeForCompilation(interpreterClass, "addYield", b -> {
+        BasicInterpreter root = parseNodeForCompilation(run, "addYield", b -> {
             b.beginRoot();
 
             b.beginReturn();
@@ -763,7 +770,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
 
     @Test
     public void testYieldInstrumentation() {
-        BasicInterpreter root = parseNodeForCompilation(interpreterClass, "addYieldInstrumented", b -> {
+        BasicInterpreter root = parseNodeForCompilation(run, "addYieldInstrumented", b -> {
             b.beginRoot();
 
             b.beginReturn();
@@ -797,7 +804,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
 
         // Instrumentation should invalidate the compiled code.
         root.getRootNodes().update(
-                        BasicInterpreterBuilder.invokeNewConfigBuilder(interpreterClass).addInstrumentation(BasicInterpreter.IncrementValue.class).build());
+                        run.bytecode().newConfigBuilder().addInstrumentation(BasicInterpreter.IncrementValue.class).build());
         assertNotCompiled(target);
         assertNotCompiled(continuationCallTarget);
 
@@ -815,7 +822,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
     @Test
     public void testCompiledSourceInfo() {
         Source s = Source.newBuilder("test", "return sourcePosition", "compiledSourceInfo").build();
-        BasicInterpreter root = parseNodeForCompilation(interpreterClass, "compiledSourceInfo", b -> {
+        BasicInterpreter root = parseNodeForCompilation(run, "compiledSourceInfo", b -> {
             b.beginSource(s);
             b.beginSourceSection(0, 21);
             b.beginRoot();
@@ -863,7 +870,7 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
 
     @Test
     public void testTagInstrumentation() {
-        BasicInterpreter root = parseNodeForCompilation(interpreterClass, "tagInstrumentation", b -> {
+        BasicInterpreter root = parseNodeForCompilation(run, "tagInstrumentation", b -> {
             b.beginRoot();
 
             // i = 0
@@ -990,8 +997,9 @@ public class BytecodeDSLCompilationTest extends TestWithSynchronousCompiling {
         return c;
     }
 
-    private static <T extends BasicInterpreterBuilder> BasicInterpreter parseNodeForCompilation(Class<? extends BasicInterpreter> interpreterClass, String rootName, BytecodeParser<T> builder) {
-        BasicInterpreter result = parseNode(interpreterClass, BytecodeDSLTestLanguage.REF.get(null), false, rootName, builder);
+    private static BasicInterpreter parseNodeForCompilation(TestRun run,
+                    String rootName, BytecodeParser<BasicInterpreterBuilder> builder) {
+        BasicInterpreter result = parseNode(run, BytecodeDSLTestLanguage.REF.get(null), rootName, builder);
         result.getBytecodeNode().setUncachedThreshold(0); // force interpreter to skip tier 0
         return result;
     }