+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/truffle/test/BytecodeDSLPartialEvaluationTest.java
@@ -26,6 +26,7 @@ package jdk.graal.compiler.truffle.test;
 
 import static com.oracle.truffle.api.bytecode.test.basic_interpreter.AbstractBasicInterpreterTest.parseNode;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Supplier;
 
@@ -43,6 +44,8 @@ import com.oracle.truffle.api.bytecode.test.BytecodeDSLTestLanguage;
 import com.oracle.truffle.api.bytecode.test.basic_interpreter.AbstractBasicInterpreterTest;
 import com.oracle.truffle.api.bytecode.test.basic_interpreter.BasicInterpreter;
 import com.oracle.truffle.api.bytecode.test.basic_interpreter.BasicInterpreterBuilder;
+import com.oracle.truffle.api.bytecode.test.basic_interpreter.AbstractBasicInterpreterTest.TestRun;
+import com.oracle.truffle.api.bytecode.test.basic_interpreter.BasicInterpreterBuilder.BytecodeVariant;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.instrumentation.EventContext;
 import com.oracle.truffle.api.instrumentation.ExecutionEventListener;
@@ -58,17 +61,21 @@ public class BytecodeDSLPartialEvaluationTest extends PartialEvaluationTest {
     protected static final BytecodeDSLTestLanguage LANGUAGE = null;
 
     @Parameters(name = "{0}")
-    public static List<Class<? extends BasicInterpreter>> getInterpreterClasses() {
-        return AbstractBasicInterpreterTest.allInterpreters();
+    public static List<TestRun> getParameters() {
+        List<TestRun> result = new ArrayList<>();
+        for (BytecodeVariant bc : AbstractBasicInterpreterTest.allVariants()) {
+            result.add(new TestRun(bc, false, false));
+        }
+        return result;
     }
 
-    @Parameter(0) public Class<? extends BasicInterpreter> interpreterClass;
+    @Parameter(0) public TestRun run;
 
     @Test
     public void testAddTwoConstants() {
         // return 20 + 22;
 
-        BasicInterpreter root = parseNodeForPE(interpreterClass, "addTwoConstants", b -> {
+        BasicInterpreter root = parseNodeForPE(run, "addTwoConstants", b -> {
             b.beginRoot();
 
             b.beginReturn();
@@ -88,7 +95,7 @@ public class BytecodeDSLPartialEvaluationTest extends PartialEvaluationTest {
     public void testAddThreeConstants() {
         // return 40 + 22 + - 20;
 
-        BasicInterpreter root = parseNodeForPE(interpreterClass, "addThreeConstants", b -> {
+        BasicInterpreter root = parseNodeForPE(run, "addThreeConstants", b -> {
             b.beginRoot();
 
             b.beginReturn();
@@ -115,7 +122,7 @@ public class BytecodeDSLPartialEvaluationTest extends PartialEvaluationTest {
     public void testAddThreeConstantsWithConstantOperands() {
         // return 40 + 22 + - 20;
 
-        BasicInterpreter root = parseNodeForPE(interpreterClass, "addThreeConstantsWithConstantOperands", b -> {
+        BasicInterpreter root = parseNodeForPE(run, "addThreeConstantsWithConstantOperands", b -> {
             b.beginRoot();
 
             b.beginReturn();
@@ -147,7 +154,7 @@ public class BytecodeDSLPartialEvaluationTest extends PartialEvaluationTest {
 
         long endValue = 10L;
 
-        BasicInterpreter root = parseNodeForPE(interpreterClass, "sum", b -> {
+        BasicInterpreter root = parseNodeForPE(run, "sum", b -> {
             b.beginRoot();
 
             BytecodeLocal i = b.createLocal();
@@ -208,7 +215,7 @@ public class BytecodeDSLPartialEvaluationTest extends PartialEvaluationTest {
         // return 3;
         // @formatter:on
 
-        BasicInterpreter root = parseNodeForPE(interpreterClass, "sum", b -> {
+        BasicInterpreter root = parseNodeForPE(run, "sum", b -> {
             b.beginRoot();
 
             b.beginTryCatch();
@@ -256,7 +263,7 @@ public class BytecodeDSLPartialEvaluationTest extends PartialEvaluationTest {
         // return 42;
         // @formatter:on
 
-        BasicInterpreter root = parseNodeForPE(interpreterClass, "sum", b -> {
+        BasicInterpreter root = parseNodeForPE(run, "sum", b -> {
             b.beginRoot();
 
             b.beginTryCatch();
@@ -306,7 +313,7 @@ public class BytecodeDSLPartialEvaluationTest extends PartialEvaluationTest {
     public void testConditionalTrue() {
         // return true ? 42 : 21;
 
-        BasicInterpreter root = parseNodeForPE(interpreterClass, "conditionalTrue", b -> {
+        BasicInterpreter root = parseNodeForPE(run, "conditionalTrue", b -> {
             b.beginRoot();
             b.beginReturn();
             b.beginConditional();
@@ -328,7 +335,7 @@ public class BytecodeDSLPartialEvaluationTest extends PartialEvaluationTest {
     public void testConditionalFalse() {
         // return false ? 21 : 42;
 
-        BasicInterpreter root = parseNodeForPE(interpreterClass, "conditionalFalse", b -> {
+        BasicInterpreter root = parseNodeForPE(run, "conditionalFalse", b -> {
             b.beginRoot();
 
             b.beginReturn();
@@ -354,7 +361,7 @@ public class BytecodeDSLPartialEvaluationTest extends PartialEvaluationTest {
         // earlyReturn(42)  // throws exception caught by intercept hook
         // return 123
         // @formatter:on
-        BasicInterpreter root = parseNodeForPE(interpreterClass, "earlyReturn", b -> {
+        BasicInterpreter root = parseNodeForPE(run, "earlyReturn", b -> {
             b.beginRoot();
             b.beginBlock();
 
@@ -379,7 +386,7 @@ public class BytecodeDSLPartialEvaluationTest extends PartialEvaluationTest {
 
         // Note: the variadic array length is not PE constant beyond 8 arguments.
         final int numVariadic = 8;
-        BasicInterpreter root = parseNodeForPE(interpreterClass, "variadicLength", b -> {
+        BasicInterpreter root = parseNodeForPE(run, "variadicLength", b -> {
             b.beginRoot();
             b.beginBlock();
 
@@ -405,7 +412,7 @@ public class BytecodeDSLPartialEvaluationTest extends PartialEvaluationTest {
         try (Context c = Context.create()) {
             c.enter();
 
-            BasicInterpreter root = parseNodeForPE(interpreterClass, "testEmptyTagInstrumentation", b -> {
+            BasicInterpreter root = parseNodeForPE(run, "testEmptyTagInstrumentation", b -> {
                 b.beginRoot();
 
                 b.beginTag(ExpressionTag.class);
@@ -440,7 +447,7 @@ public class BytecodeDSLPartialEvaluationTest extends PartialEvaluationTest {
 
             String text = "return 20 + 22";
             Source s = Source.newBuilder("test", text, "testUnwindTagInstrumentation").build();
-            BasicInterpreter root = parseNodeForPE(BytecodeDSLTestLanguage.REF.get(null), interpreterClass, "testUnwindTagInstrumentation", b -> {
+            BasicInterpreter root = parseNodeForPE(BytecodeDSLTestLanguage.REF.get(null), run, "testUnwindTagInstrumentation", b -> {
                 b.beginSource(s);
                 b.beginSourceSection(0, text.length());
                 b.beginRoot();
@@ -507,13 +514,15 @@ public class BytecodeDSLPartialEvaluationTest extends PartialEvaluationTest {
         return () -> result;
     }
 
-    private static <T extends BasicInterpreterBuilder> BasicInterpreter parseNodeForPE(Class<? extends BasicInterpreter> interpreterClass, String rootName, BytecodeParser<T> builder) {
-        return parseNodeForPE(LANGUAGE, interpreterClass, rootName, builder);
+    private static BasicInterpreter parseNodeForPE(TestRun run,
+                    String rootName, BytecodeParser<BasicInterpreterBuilder> builder) {
+        return parseNodeForPE(LANGUAGE, run, rootName, builder);
     }
 
-    private static <T extends BasicInterpreterBuilder> BasicInterpreter parseNodeForPE(BytecodeDSLTestLanguage language, Class<? extends BasicInterpreter> interpreterClass, String rootName,
-                    BytecodeParser<T> builder) {
-        BasicInterpreter result = parseNode(interpreterClass, language, false, rootName, builder);
+    private static <T extends BasicInterpreterBuilder> BasicInterpreter parseNodeForPE(BytecodeDSLTestLanguage language,
+                    TestRun run, String rootName,
+                    BytecodeParser<BasicInterpreterBuilder> builder) {
+        BasicInterpreter result = parseNode(run, language, rootName, builder);
         result.getBytecodeNode().setUncachedThreshold(0); // force interpreter to skip tier 0
         return result;
     }