+++ truffle/src/com.oracle.truffle.api.bytecode.test/src/com/oracle/truffle/api/bytecode/test/basic_interpreter/BranchTest.java
@@ -412,5 +412,4 @@ public class BranchTest extends AbstractBasicInterpreterTest {
 
         assertEquals(42L, root.call());
     }
-
 }