+++ truffle/src/com.oracle.truffle.api.bytecode.test/src/com/oracle/truffle/api/bytecode/test/basic_interpreter/LocalsTest.java
@@ -373,7 +373,7 @@ public class LocalsTest extends AbstractBasicInterpreterTest {
         List<LocalVariable> locals = b.getLocals();
         assertEquals(5, locals.size());
         assertEquals(42L, root.getCallTarget().call());
-        if (run.hasBlockScoping()) {
+        if (run.hasBlockScoping() && !run.testTracer()) {
             assertEquals(0, locals.get(0).getStartIndex());
             assertEquals(endBci, locals.get(0).getEndIndex());
             assertEquals("l0", locals.get(0).getName());
@@ -804,7 +804,7 @@ public class LocalsTest extends AbstractBasicInterpreterTest {
 
     }
 
-    private <T extends BasicInterpreterBuilder> void assertParseFailure(BytecodeParser<T> parser) {
+    private void assertParseFailure(BytecodeParser<BasicInterpreterBuilder> parser) {
         assertThrows(IllegalArgumentException.class, () -> parseNode("invalid", parser));
     }